Fix typo in `MockAdapter` (v2.x) (#766)
## Description

This PR fixes a typo in the imported `MockAdapter` from
`axios-mock-adapter`.

Same as #765, but for the `v2.x` branch.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.