Fix typo in `MockAdapter` (#765)
## Description

This PR fixes a typo in the imported `MockAdapter` from
`axios-mock-adapter`.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.