Rename User interface to DirectoryUser (#774)
## Description

* This is a breaking change to rename the User interface, which is
taking up a namespace that we would otherwise like to use, to
DirectoryUser. This aligns the interface with the naming used elsewhere
across the WorkOS platform.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[x] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.