Rename Group interface to DirectoryGroup (#775)
## Description

* This is a breaking change to rename the Group interface. While not
strictly necessary as the User -> DirectoryUser rename was, this is done
to keep consistency across the SDK and to closer align the Node SDK
interfaces with other SDKs and with the rest of our platform.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[x] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.