Rename DeserializedList to List (#800)
## Description

* Had previously inverted the naming structure of the serialization of
List to help make an easier migration. Now that the bulk of the work of
migrating to camelCase is complete we're making the swap to keep the
naming in line with the rest of the types.
* DeseralizedList is becoming List and List is becoming ListResponse.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.