Update Session interfaces to match API (#802)
## Description

workos/workos#22251 updated the shape of the session response. This
updates the SDK to match it.

I'll update the docs as well to have the same shape as in this PR.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[x] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.