Ensure the SDK version in code matches `package.json` (#809)
## Description

This PR adds a test that ensures that the SDK version number defined in
the code is the same as the one in `package.json`.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.