Export `Session` types (#810)
## Description

This PR exports the `Session` and its related types so they're
accessible to consumers of the SDK.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.