Make password optional when creating users (#814)
## Description

Password is now optional when creating users in the API as of
workos/workos#22384. This matches it.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.