Make `session` required on `AuthenticationResponse` (#811)
## Description

This PR fixes a bug in the `AuthenticationResponse` type where the
`session` field was marked as possibly being `undefined`.

The `session` will always be present on the `AuthenticationResponse`.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.

Co-authored-by: Will Duffy <willman@workos.com>