Remove `--tag=next` in npm publish (#816)
## Description

This PR removes the `--tag=next` so that we can publish to the `latest`
tag.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.