Properly serialize args for listConnections and listDirectories (#822)
## Description

* When we added serialization, we introduced a bug for these two
endpoints that did not properly pass through the serialization of
arguments for these two endpoints. This resolves the issue and adds
tests to check for regressions in the future.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.