Rename `authenticateUserWithToken` to `authenticateUserWithCode` (#820)
## Description

This PR renames the `authenticateUserWithToken` method to
`authenticateUserWithCode` to better reflect its usage.

This is a breaking change to the interface of the beta User Management
APIs.

Resolves USRLD-799.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.