Remove distinction between "Managed" and "Unmanaged" users (#831)
## Description

This PR removes the distinction between "Managed" and "Unmanaged" users
from the SDK.

This concept is being removed from the API in the near future.

This is a breaking change to the User Management portion of the SDK.

Resolves USRLD-843.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.