Migrate completeEmailVerification to verifyEmail (#837)
## Description

* Changes the completeEmailVerification path and body to match the
latest updates from the API. It now not only takes a token (now renamed
to code) as an argument, but also requires the userId of the user.
* Fixes USRLD-849, USRLD-934

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.

---------

Co-authored-by: Marshall Bowers <elliott.codes@gmail.com>