Add rate limiting to pagination (#866)
## Description

Our auto pagination doesn't currently respect a limit of 4 RPS (the
limit on our `list_users` endpoint). This is a proposed fix, though it
applies to all endpoints.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.

---------

Co-authored-by: Ash Godfrey <ashgodfrey@Ashs-MacBook-Pro.local>