Update main from 3.x (#891)
## Description
Updates `main` from the `3.x` branch, which we're using for minor/patch
releases while we work on the next major release on `main`.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.