Split out with and without secrets factors (#929)
## Description
Enrolling a TOTP auth factor always returns the `qrCode`, `secret`, and
`uri` in the response, but retrieving and listing auth factors do not.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.



#### PR Dependency Tree


* **PR #929** 👈

This tree was auto-generated by
[Charcoal](https://github.com/danerwilliams/charcoal)

---------

Co-authored-by: Amy Hanlon <amy@workos.com>