v5.0.0  (#934)
## Description

Updates version specifiers to `5.0.0` in preparation of creating a new
`v5.0.0` tag and release.

Note that I'm bumping this to `v5.0.0` since `main` appears to currently
contain a breaking change in the form of a renamed parameter:
https://github.com/workos/workos-node/commit/0f0bab76be4c6a20cc5161da55edc5886afcf982

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.