Deploy new versions to NPM when a new release is cut in GitHub (#997)
## Description
This migrates the deploy step in Semaphore to also run in GitHub
actions. This is set up to run automatically when we cut a new release
(aka `git push --tags`) or you can manually trigger a release using the
`Actions` tab.

This PR is set to only `--dry-run` the actual push because I can't
actually test it until it's merged.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.

---------

Co-authored-by: Amy Hanlon <amy@workos.com>