Finalize the move to GitHub Actions for CI and Releases (#1003)
## Description
With this PR, Semaphore is fully removed from the repo and all release
handling is done via GitHub Actions as well.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.