Export `BadRequestException` to perform type narrowing on error handling (#1009)
## Description

Export `BadRequestException` to be able to narrow it's type for error
handling:

```ts
try {
      await this.client.userManagement.createUser(options);
    } catch (error) {
      if (error instanceof BadRequestException) {}
}}
```

As an example, `createUser` returns two exceptions that are inherited
from our `BadRequestException` class.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.

Co-authored-by: Paul Asjes <108872335+PaulAsjes@users.noreply.github.com>