Add `role_slug` to user management invite flow (#1008)
## Description
`role_slug` is an optional parameter when sending an invite

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.

---------

Co-authored-by: Matt Dzwonczyk <9063128+mattgd@users.noreply.github.com>