Add 'password_hash' and 'password_hash_type' to User creation (#1014)
## Description
This PR ensures that the User create process allows for the
`password_hash` and `password_hash_type` fields, in addition to the
existing `password` field, similar to the Update process.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[x] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.

https://github.com/workos/workos/pull/26060