AppInfo constructor options (#1010)
## Description
Adds an optional `appInfo` object to the constructor options. This
allows you to specify the app that's using workos-node in the
user-agent.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.