Support `organizationId` as a valid parameter for `events` (#1034)
## Description
Allow passing an `organizationId` into the options for the Events API 
## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.