Add explicit handling for rate-limit errors (#1033)
## Description

Introduces a new `RateLimitExceededException` to be thrown when a `429`
error is encountered, rather than defaulting to a more generic error
exception class.

This also allows the reading of the `Retry-After` header, which is
helpful for clients to retry, as well as debugging.

There's also some light refactoring included.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.