Add `inviterUserId` to `invitation` (#1040)
## Description

Exposes the `inviterUserId` field on the `invitation` object.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.