Pre-release deployments should use the version of the tag
npm doesn't let you publish multiple of the same version, but we also
don't want to have to update package.json for every prerelease deploy we
want to push. This change updates package.json to match the name of the
tag triggering this beta deploy ensuring our versions are always
unique.

I also reverted back to updating the `next` tag as that should always
point to the latest pre-release build.