Fix `NodeHttpClient` handling of `URLSearchParam` request bodies  (#1071)
## Description

Previously the client was passing the body directly down to the low
level `write` and `byteLength` calls, which failed in cases where the
body wasn't already serialized as a `string` or form of buffer.

This was not noticed on newer versions of Node since they include a
global `fetch`, resulting in the SDK preferring the `FetchHttpClient`.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.