Update `grant_type=refresh_token` types (#1073)
## Description

The refresh_token response used to be more limited, but that is no
longer the case.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.