Add authentication events (#998)
## Description

Adding types/parsing for authentication events - all the succeeded
events and only the failed events for password and Magic Auth. We will
add the other failed events as we complete support for them.

Resolves DAAP-913

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.

---------

Co-authored-by: Blair Lunceford <74257063+blairlunceford@users.noreply.github.com>