v7.16.0 (#1083)
## Description

- Add authentication events:
https://github.com/workos/workos-node/pull/998

I know we just cut a release today, and I wouldn't normally push up a
new one so soon, but this will unblock a customer, and needs to be
released ASAP.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.