Add `organizationId` parameter to `AuthenticateWithRefreshTokenOptions` (#1093)
## Description

Adds the new `organizationId` parameter to the
`AuthenticateWithRefreshTokenOptions`.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.