Update `refreshAndSealSessionData` to accept `organizationId` (#1100)
## Description

This matches the `authenticateWithRefreshToken` method which also
accepts an `organizationId` to allow "switching".

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.