fix: Ensure organizationId is included in SessionCookieData during sealSessionDataFromAuthenticationResponse (#1108)
## Description
They types were a slight lie, since you could have had a session with an
org id in the access token the but the organization id in the sealed
cookie data didn't include it. This required users to always decode the
jwt even in cases where they were trying to abstract away using the new
sealed session data.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.