Add certificate renewal intent for Admin Portal intents (#1104)
## Description
Adding the certificate renewal intent as an option for link generation
## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.