Rework session helpers (#1105)
## Description

Refactors the session helpers to add a new session class. The old helper
methods will stay put, but we'll likely add a deprecation comment and
remove them in a future major version.

See this
[gist](https://gist.github.com/PaulAsjes/e1f80d5451f988db2acdf6ee894fcc46)
as an example of how this would work in practise.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[x] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.

---------

Co-authored-by: Matt Dzwonczyk <9063128+mattgd@users.noreply.github.com>