Export the correct org domain interface (#1119)
## Description
We are currently exporting a version of `OrganizationDomain.interface`
that was incorrect. This fixes that and tidies up the imports.


## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.