Reject non-retryable error in node http client (#1155)
## Description

In the `NodeHttpClient`, `TypeError`s are considered retryable. However,
if we get any other errors, we should reject with the given error.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.