Remove org Stripe customer ID setter/getter (#1160)
## Description

The same thing can be accomplished with the
`organizations.updateOrganization` function.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.