Add `Passkey` to `AuthenticationMethod` union (#1166)
## Description

Adds `'Passkey'` as a possible value of `AuthenticationMethod`.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.