Add accept-invite function (#1180)
## Description

The endpoint is still behind a feature flag, but this makes it
available.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[x] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.

https://github.com/workos/workos/pull/32872