Use a non-default ttl with iron-session (#1186)
## Description
iron-session has a default ttl of 14 days, which is problematic in the
edge case where a user has set their WorkOS sessions to expire > 14
days. In that scenario, iron-session will expire first, making unsealing
impossible which means you can't refresh the session.

This fix changes the default ttl to 0, which sets it to the max of ~10
years.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.