set JWKS cache duration to 5 minutes (#1196)
## Description

By default, jose sets a `cooldownDuration` to 30 seconds for JWKS cache.
This PR changes it to 5 minutes.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.