Add device fingerprint to action context payloads (#1199)
When Radar is enabled, the device fingerprint captured by Radar is
forwarded in action request payloads. This PR adds the attribute to
action interfaces.

## Description

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.