Update Coana Guardrail to not require API key for run (#1214)
## Description
Update the Coana Guardrail to no longer require an API key to run. This
allows for forks of public repos such as this one to run successfully.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.