Add externalId support to user (#1228)
Allows adding one during create and update and includes the externalId
in the user object.

## Description

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[x] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.

^ docs are coming but looking to get this out in a preview release ahead
of time.