Vault api (#1236)
## Description

Adds SDK support for WorkOS Vault. Vault simplifies encryption and key
management for modern web applications. It provides a service for
storing sensitive data secured with scoped encryption keys, as well as a
utility to encrypt data on the client using the same keys.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[X] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.
[Docs PR](https://github.com/workos/workos/pull/35787)