Add metadata to organizations and users (#1241)
## Description

Supports passing metadata on updating and creation of both as well.

Right now the API doesn't always return metadata, but it will eventually
always return empty object if there is none, and so I've made it so that
users and organizations will always have a metadata object after
deserialization.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.