Rename Vault secrets to objects (#1247)
## Description

Renames `secrets` in Vault to `object`. The methods and interfaces for
`secrets` are aliased to the new `object` ones to help with the
transition. Eventually those will be removed completely.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[x] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.

[Docs PR](https://github.com/workos/workos/pull/37306)