General dev dependency updates (#1254)
>[!note]
>This change is against the `version-8` branch.

Update to latest of the following dev dependencies
- TypeScript
- ts-jest
- Jest
- Supertest

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.