Add organizationId to organization domain (#1259)
## Description

Add `organizationId` to the response as indicated in
work-os-playground-domain-verification-v4hbbh

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.