Deprecate `context` option on `getAuthorizationUrl` (#1263)
We no longer need applications' initiate login endpoints to return the
`context` query parameter when getting the authorization URL. This PR
deprecates that option from the `getAuthorizationUrl` and `signIn`
functions and prints a message to the console when it's used. In a
future release, we can remove this option entirely.