Export Identity/IdentityResponse interfaces (#1291)
## Description

This adds the export to the `user_management/interfaces/index.ts` barrel
file so that it will get exported for consumers of the library.

Fixes #1290

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.