Fix `GetTokenOptions[scopes]` type (#1348)
## Description

This was accidentally typed as a single-item tuple, so passing multiple
scopes currently results in a type error.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.