Add Salesforce OAuth as an authentication provider (#1361)
## Description

We are rolling out support for Salesforce OAuth. This adds it to the
list of providers.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.