add convex conditional export (#1372)
## Description

Adds a conditional export for the convex runtime.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.

Co-authored-by: Nick Nisi <nick.nisi@workos.com>