Remove dependency on pluralize (#1379)
## Description

[plurarize](https://www.npmjs.com/package/pluralize) is only used once
in our sdk and it brings almost no value. Also, the package itself was
last updated 6 years ago and does not support `esm`, so I think it's
better to remove this now, as it already creates problems with esm only
bundlers.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.