Add passkey authentication events (#1392)
## Description

A bug in our API prevented these from being returned but also we are
missing them from the types in this library. The API bug has been fixed
and this fixes the SDK.

## Documentation

Does this require changes to the WorkOS Docs? E.g. the [API
Reference](https://workos.com/docs/reference) or code snippets need
updates.

```
[ ] Yes
```

If yes, link a related docs PR and add a docs maintainer as a reviewer.
Their approval is required.